+++ app/database/migrations/2025_03_19_223715_remove_logo_column_vehicles.php
@@ -6,23 +6,28 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
-    public function up(): void
+    public function up()
     {
-        Schema::table('vehicles', function (Blueprint $table) {
-            $table->dropColumn('logo');
-        });
+        // Verifica se la colonna 'logo' esiste già
+        if (!Schema::hasColumn('vehicles', 'logo')) {
+            // Se la colonna non esiste, la aggiungi
+            Schema::table('vehicles', function (Blueprint $table) {
+                $table->string('logo')->nullable()->after('model_id'); // Aggiungi la colonna logo
+            });
+        } else {
+            // Ignora la migrazione se la colonna logo esiste già
+            // Non facciamo nulla
+            return;
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
-    public function down(): void
+    public function down()
     {
         Schema::table('vehicles', function (Blueprint $table) {
-            $table->string('logo')->nullable();
+            // Se la colonna logo esiste, la rimuoviamo
+            if (Schema::hasColumn('vehicles', 'logo')) {
+                $table->dropColumn('logo');
+            }
         });
     }
 };
